# Removes MessageAttribute in favor of properties on Message class