--- a/src/Renci.SshNet/Messages/Authentication/InformationResponseMessage.cs
+++ b/src/Renci.SshNet/Messages/Authentication/InformationResponseMessage.cs
@@ -6,9 +6,26 @@ namespace Renci.SshNet.Messages.Authentication
     /// <summary>
     /// Represents SSH_MSG_USERAUTH_INFO_RESPONSE message.
     /// </summary>
-    [Message("SSH_MSG_USERAUTH_INFO_RESPONSE", 61)]
     internal sealed class InformationResponseMessage : Message
     {
+        /// <inheritdoc />
+        public override string MessageName
+        {
+            get
+            {
+                return "SSH_MSG_USERAUTH_INFO_RESPONSE";
+            }
+        }
+
+        /// <inheritdoc />
+        public override byte MessageNumber
+        {
+            get
+            {
+                return 61;
+            }
+        }
+
         /// <summary>
         /// Gets authentication responses.
         /// </summary>