--- a/src/Renci.SshNet/Messages/Connection/ChannelWindowAdjustMessage.cs
+++ b/src/Renci.SshNet/Messages/Connection/ChannelWindowAdjustMessage.cs
@@ -3,9 +3,26 @@
     /// <summary>
     /// Represents SSH_MSG_CHANNEL_SUCCESS message.
     /// </summary>
-    [Message("SSH_MSG_CHANNEL_WINDOW_ADJUST", 93)]
     public class ChannelWindowAdjustMessage : ChannelMessage
     {
+        /// <inheritdoc />
+        public override string MessageName
+        {
+            get
+            {
+                return "SSH_MSG_CHANNEL_WINDOW_ADJUST";
+            }
+        }
+
+        /// <inheritdoc />
+        public override byte MessageNumber
+        {
+            get
+            {
+                return 93;
+            }
+        }
+
         /// <summary>
         /// Gets number of bytes to add to the window.
         /// </summary>