--- a/src/Renci.SshNet/Messages/Transport/DebugMessage.cs
+++ b/src/Renci.SshNet/Messages/Transport/DebugMessage.cs
@@ -3,11 +3,28 @@
     /// <summary>
     /// Represents SSH_MSG_DEBUG message.
     /// </summary>
-    [Message("SSH_MSG_DEBUG", 4)]
     public class DebugMessage : Message
     {
         private byte[] _message;
         private byte[] _language;
+
+        /// <inheritdoc />
+        public override string MessageName
+        {
+            get
+            {
+                return "SSH_MSG_DEBUG";
+            }
+        }
+
+        /// <inheritdoc />
+        public override byte MessageNumber
+        {
+            get
+            {
+                return 4;
+            }
+        }
 
         /// <summary>
         /// Gets a value indicating whether the message to be always displayed.