… src/Renci.SshNet/Messages/MessageAttribute.cs
… src/Renci.SshNet/Messages/Transport/NewKeysMessage.cs
… src/Renci.SshNet/Messages/Transport/ServiceAcceptMessage.cs
… src/Renci.SshNet/Messages/Transport/ServiceRequestMessage.cs
… src/Renci.SshNet/Messages/Transport/UnimplementedMessage.cs
… test/Renci.SshNet.Benchmarks/Messages/MessageBenchmarks.cs
… test/Renci.SshNet.Benchmarks/Renci.SshNet.Benchmarks.csproj
… test/Renci.SshNet.IntegrationTests/Properties/AssemblyInfo.cs
… test/Renci.SshNet.Tests/Classes/Messages/Connection/ChannelDataMessageTest.cs
… test/Renci.SshNet.Tests/Classes/Messages/Connection/ChannelOpen/ChannelOpenMessageTest.cs
… test/Renci.SshNet.Tests/Classes/Messages/Transport/IgnoreMessageTest.cs
… test/Renci.SshNet.Tests/Classes/Messages/Transport/KeyExchangeDhGroupExchangeGroupBuilder.cs
… test/Renci.SshNet.Tests/Classes/Messages/Transport/KeyExchangeDhGroupExchangeReplyBuilder.cs
… test/Renci.SshNet.Tests/Classes/Messages/Transport/KeyExchangeDhGroupExchangeRequestTest.cs
… test/Renci.SshNet.Tests/Classes/SessionTest_ConnectedBase.cs
… test/Renci.SshNet.Tests/Classes/SessionTest_Connected_ServerAndClientDisconnectRace.cs
… test/Renci.SshNet.Tests/Properties/AssemblyInfo.cs